ccw: Fix requirements for processing adapter interrupts.
We currently mandate that the driver clears the summary indicator
before processing the queue indicator; this is bogus, as the requirement
for interrupt avoidance is rather that the driver unsets the summary
indicator before before it stops looking at the queue indicator.

In fact, the best way to get a race-free implementation of the interrupt
handler is to process the queue indicators twice; let's add a recommondation
to do that.

Signed-off-by: Cornelia Huck <cornelia.huck@de.ibm.com>
Signed-off-by: Rusty Russell <rusty@au1.ibm.com> (merged)