makediff: work around for lstlisting
latexdiff seems to go wild trying to diff lstlisting.
But marking lstlisting as picture - so old text is
commented out - then removing the comment sign -
seems to work, with the only disadvantage being
that listing is diffed as a whole.
Seems to be the best I can come up with ATM.

Signed-off-by: Michael S. Tsirkin <mst@redhat.com>