VIRTIO-137: used ring: define the meaning and requirements of the len field.
We said what it was for, and noted why.  We didn't place any requirements
on it, nor clearly spell out the implications of its use.

This clarification comes particularly from noticing that QEMU didn't
set len correctly, and philosophising over the correct value when
an error has occurred.

(Wording precision feedback from Michael and Cornelia - Thanks!)

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>

Signed-off-by: Michael S. Tsirkin <mst@redhat.com>